+++ PostProcess/netcdf_npp_fns.py
@@ -17,7 +17,7 @@ from os.path import normpath, isfile, join
 from os import remove
 import time
 from netCDF4 import Dataset
-from numpy import arange, float32
+from numpy import arange, float64
 
 from nc_low_level_fns import generate_mnthly_atimes
 
@@ -63,8 +63,8 @@ def create_npp_ncs(form, metrics, crop_names, land_use, out_dir, study, npp_unit
 
     # build lat long arrays
     # =====================
-    alons = arange(ll_lon, ur_lon, resol, dtype=float32)
-    alats = arange(ll_lat, ur_lat, resol, dtype=float32)
+    alons = arange(ll_lon, ur_lon, resol, dtype=float64)
+    alats = arange(ll_lat, ur_lat, resol, dtype=float64)
     num_alons = len(alons)
     num_alats = len(alats)
     form.bbox_nc = bbox
@@ -133,7 +133,7 @@ def create_npp_ncs(form, metrics, crop_names, land_use, out_dir, study, npp_unit
         # create the variable (4 byte float in this case)
         # to create a netCDF variable, use the createVariable method of a Dataset (or Group) instance.
         # first argument is name of the variable, second is datatype, third is a tuple with the name (s) of the dimension(s).
-        # lats = nc_dset.createVariable('latitude',dtype('float32').char,('latitude',))
+        # lats = nc_dset.createVariable('latitude',dtype('float64').char,('latitude',))
         #
         lats = nc_dset.createVariable('latitude','f4',('latitude',))
         lats.description = 'degrees of latitude North to South in ' + str(resol) + ' degree steps'
